requirements.txt
flask
flask-cors
numpy
pillow
face-recognition
setuptools
git+https://github.com/ageitgey/face_recognition_models
